requirements.txt
Flask==1.0.2
psycopg2-binary==2.8.6
requests==2.19.1
python-dotenv==0.13.0
itsdangerous==2.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
Jinja2==3.0.3
MarkupSafe==2.1.1
pycparser==2.21
six==1.16.0
Werkzeug==2.0.3